refactor: independent def_maps for blocks
Blocks play a special role in hir_def because they are a scope and a
stmt at the same time. To avoid coupeling the body and the def_map (and
for more reliable behaviour) Each blog now has an independent query for
its independent def_map

+++ crates/syntax/src/ast/expr_ext.rs
@@ -225,7 +225,7 @@ impl ast::SiRealNumber {
 }
 
 impl ast::IntNumber {
-    pub fn value(&self) -> i64 {
+    pub fn value(&self) -> i32 {
         self.syntax.text().parse().unwrap()
     }
 }
